plugins: repo: add routines to fetch plugin archive
